% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFunction.R
\name{getSum}
\alias{getSum}
\title{Sum variables from fieldbook}
\usage{
getSum(var1, var2)
}
\arguments{
\item{var1}{numeric vector variable 1 from fb}

\item{var2}{numeric vector variable 2 from fb}
}
\description{
Sum of two numeric vectors
}
\details{
nombre > tipo > description
}
\author{
Alonso Medina
}
